requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.13.4
    # via html-sanitizer
boto3==1.38.33
    # via
    #   eodhp-utils
    #   harvest-transformer (pyproject.toml)
botocore==1.38.33
    # via
    #   boto3
    #   eodhp-utils
    #   harvest-transformer (pyproject.toml)
    #   s3transfer
certifi==2025.4.26
    # via
    #   pulsar-client
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via harvest-transformer (pyproject.toml)
eodhp-utils @ git+https://github.com/EO-DataHub/eodhp-utils@v0.1.12
    # via harvest-transformer (pyproject.toml)
faker==37.3.0
    # via eodhp-utils
html-sanitizer==2.5.0
    # via harvest-transformer (pyproject.toml)
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via opentelemetry-api
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.33
    # via harvest-transformer (pyproject.toml)
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.24.0
    # via
    #   eodhp-utils
    #   harvest-transformer (pyproject.toml)
jsonschema-specifications==2025.4.1
    # via jsonschema
lxml==5.4.0
    # via
    #   html-sanitizer
    #   lxml-html-clean
lxml-html-clean==0.4.2
    # via html-sanitizer
opentelemetry-api==1.34.1
    # via
    #   eodhp-utils
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-processor-baggage
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-instrumentation==0.55b1
    # via opentelemetry-instrumentation-logging
opentelemetry-instrumentation-logging==0.55b1
    # via eodhp-utils
opentelemetry-processor-baggage==0.55b1
    # via eodhp-utils
opentelemetry-sdk==1.34.1
    # via
    #   eodhp-utils
    #   opentelemetry-processor-baggage
opentelemetry-semantic-conventions==0.55b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==25.0
    # via opentelemetry-instrumentation
pulsar-client==3.7.0
    # via eodhp-utils
pysubnettree==0.37
    # via eodhp-utils
python-dateutil==2.9.0.post0
    # via botocore
python-json-logger==3.3.0
    # via eodhp-utils
pyyaml==6.0.2
    # via harvest-transformer (pyproject.toml)
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   eodhp-utils
    #   harvest-transformer (pyproject.toml)
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
s3transfer==0.13.0
    # via boto3
six==1.17.0
    # via python-dateutil
soupsieve==2.7
    # via beautifulsoup4
typing-extensions==4.14.0
    # via
    #   beautifulsoup4
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   referencing
tzdata==2025.2
    # via faker
urllib3==2.4.0
    # via
    #   botocore
    #   requests
wrapt==1.17.2
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-processor-baggage
zipp==3.23.0
    # via importlib-metadata
